Refactored theme types
