+++ src/components/ui/scroll-view/scroll-view.styles.tsx
@@ -1,4 +1,4 @@
-import { ThemeVariables } from '@/config/theme/theme';
+import { ThemeVariables } from '@/types/theme';
 import { StyleSheet } from 'react-native';
 
 const styles = (theme: ThemeVariables) =>
